+++ packages/rxget/lib/src/get_core/get_core.dart
@@ -5,4 +5,3 @@ export 'src/get_interface.dart';
 export 'src/get_main.dart';
 export 'src/log.dart';
 export 'src/smart_management.dart';
-export 'src/typedefs.dart';