+++ packages/rxget/lib/src/get_core/src/typedefs.dart
@@ -1,2 +1,0 @@
-/// A callback that returns an updated value of type [T].
-typedef ValueUpdater<T> = T Function();