+++ packages/rxget/lib/src/get_rx/src/rx_types/rx_core/rx_interface.dart
@@ -16,23 +16,3 @@ abstract class RxInterface<T> implements ValueListenable<T> {
     bool? cancelOnError,
   });
 }
-
-/// An error message class shown when [Obx] or [GetX] widgets are used improperly.
-///
-/// This typically occurs when observable variables are not placed within
-/// the correct reactive scope.
-class ObxError {
-  /// Creates a const [ObxError].
-  const ObxError();
-  @override
-  String toString() {
-    return """
-      [Get] the improper use of a GetX has been detected. 
-      You should only use GetX or Obx for the specific widget that will be updated.
-      If you are seeing this error, you probably did not insert any observable variables into GetX/Obx 
-      or insert them outside the scope that GetX considers suitable for an update 
-      (example: GetX => HeavyWidget => variableObservable).
-      If you need to update a parent widget and a child widget, wrap each one in an Obx/GetX.
-      """;
-  }
-}