+++ packages/rxget/lib/src/get_state_manager/get_state_manager.dart
@@ -5,5 +5,5 @@ export 'src/rx_flutter/rx_obl_widget.dart';
 export 'src/rx_flutter/rx_obx_widget.dart';
 export 'src/simple/get_controllers.dart';
 export 'src/simple/get_state.dart';
-export 'src/simple/list_notifier.dart' show RxLifecycleDebug;
+export 'src/simple/list_notifier.dart' show ObxError, RxLifecycleDebug;
 export 'src/simple/simple_builder.dart';